CLN,TYP: Use subsitutions in info docstrings (#33673)
* add subs in info.py

* substitute docstring in frame.py

* Add type annotations

* reinstate data parameter

* revert removed line

* indent docstring

* fix validation of docstrings

* fix indentation of max_columns
